+++ Election/District.pde
@@ -1,7 +1,7 @@
 class District {
   State state;
   String number;
-  HashMap<Candidate, Integer> candidates;
+  HashMap<Candidate, Float> candidates;
   PShape district;
   PShape country;
   String stateCode;
@@ -11,7 +11,7 @@ class District {
   District(State parentState, String districtNo, PShape map){
     state = parentState;
     number = districtNo;
-    candidates = new HashMap<Candidate, Integer>();
+    candidates = new HashMap<Candidate, Float>();
     country = map;
     districtColor = color(0, 0, 0);
     
@@ -31,7 +31,7 @@ class District {
   }
     
    Candidate getWinner(int year) {
-    int maxValue = 0;
+    Float maxValue = 0.0;
     Candidate winner = new Candidate("a, a","R","a");
     for(Candidate candidate : candidates.keySet()) {
      if(this.candidates.get(candidate) >= maxValue) {
